% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mhqol.R
\name{shiny_mhqol}
\alias{shiny_mhqol}
\title{Launch Shiny MHQoL Interface}
\usage{
shiny_mhqol(display.mode = "normal")
}
\arguments{
\item{display.mode}{The display mode to be passed to \link[shiny]{runApp}. Default is "normal".}
}
\value{
NULL (launches the Shiny app).
}
\description{
\code{shiny_mhqol} launches a Shiny interface for browser-based MHQoL calculations.
}
\details{
This function starts the Shiny app for Mental Health Quality of Life (MHQoL) scoring,
allowing users to interactively input data and calculate utility scores.
}
\examples{
if(interactive()) {
  shiny_mhqol()
  shiny_mhqol(display.mode = "normal")
}
}
